Refactor directory structure to comply with clean architecture
