+++ src/movies/entrypoints/flask_app.py
@@ -1,7 +1,7 @@
 from flask import Flask, request, jsonify
 import json
-from movies.category import Category
-from movies.sql_alchemy_repository import SqlAlchemyRepository
+from movies.domain.category import Category
+from movies.adapters.sql_alchemy_repository import SqlAlchemyRepository
 from movies.utils.invalid_api_usage import InvalidAPIUsage
 from movies.recommendations.recommender_factory import RecommenderFactory, AlgorithmType
